cloud-function-http/requirements.txt
functions-framework==3.5.0
google-cloud-storage==2.14.0
mysql-connector-python==8.3.0
